Refactor FCI logic to use RecursiveDiscriminatingPathRule.
Replaced R0R4StrategyTestBased with RecursiveDiscriminatingPathRule to streamline FCI logic and improve clarity. Centralized recursive blocking logic, removed redundancies, and updated all references across the codebase.
